feat: refactor Equal Pay UI components and calculations for improved clarity and functionality
